+++ SerenaDeploy-agent/src/main/java/com/serena/FutureBasedBuildProcess.java
@@ -1,55 +1,45 @@
 package com.serena;
 
-import java.util.concurrent.*;
-import jetbrains.buildServer.*;
-import jetbrains.buildServer.agent.*;
-import org.jetbrains.annotations.*;
+import jetbrains.buildServer.RunBuildException;
+import jetbrains.buildServer.agent.BuildFinishedStatus;
+import jetbrains.buildServer.agent.BuildProcess;
+import org.jetbrains.annotations.NotNull;
 
-public abstract class FutureBasedBuildProcess implements BuildProcess, Callable<BuildFinishedStatus>
-{
+import java.util.concurrent.*;
+
+public abstract class FutureBasedBuildProcess implements BuildProcess, Callable<BuildFinishedStatus> {
     private Future<BuildFinishedStatus> myFuture;
 
-    public void start() throws RunBuildException
-    {
-        try
-        {
+    public void start() throws RunBuildException {
+        try {
             myFuture = Executors.newSingleThreadExecutor().submit(this);
-        } catch (final RejectedExecutionException e)
-        {
+        } catch (final RejectedExecutionException e) {
             throw new RunBuildException(e);
         }
     }
 
-    public boolean isInterrupted()
-    {
+    public boolean isInterrupted() {
         return myFuture.isCancelled() && isFinished();
     }
 
-    public boolean isFinished()
-    {
+    public boolean isFinished() {
         return myFuture.isDone();
     }
 
-    public void interrupt()
-    {
+    public void interrupt() {
         myFuture.cancel(true);
     }
 
     @NotNull
-    public BuildFinishedStatus waitFor() throws RunBuildException
-    {
-        try
-        {
+    public BuildFinishedStatus waitFor() throws RunBuildException {
+        try {
             final BuildFinishedStatus status = myFuture.get();
             return status;
-        } catch (final InterruptedException e)
-        {
+        } catch (final InterruptedException e) {
             throw new RunBuildException(e);
-        } catch (final ExecutionException e)
-        {
+        } catch (final ExecutionException e) {
             throw new RunBuildException(e);
-        } catch (final CancellationException e)
-        {
+        } catch (final CancellationException e) {
             return BuildFinishedStatus.INTERRUPTED;
         }
     }