+++ SerenaDeploy-agent/src/main/java/com/serena/SRAHelper.java
@@ -1,8 +1,10 @@
 package com.serena;
 
+import com.intellij.openapi.diagnostic.Logger;
 import com.urbancode.commons.util.https.OpenSSLProtocolSocketFactory;
 import com.urbancode.vfs.common.ClientChangeSet;
 import com.urbancode.vfs.common.ClientPathEntry;
+import jetbrains.buildServer.agent.BuildProgressLogger;
 import org.apache.commons.httpclient.HttpClient;
 import org.apache.commons.httpclient.HttpMethodBase;
 import org.apache.commons.httpclient.UsernamePasswordCredentials;
@@ -17,25 +19,24 @@ import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONObject;
 
 import javax.ws.rs.core.UriBuilder;
-import java.lang.reflect.Array;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import jetbrains.buildServer.agent.BuildProgressLogger;
 
-
-public class SRAHelper
-{
+public class SRAHelper {
+    private static final Logger logger = Logger.getInstance(SRAHelper.class.getName());
 
     String sraUrl;
     String username;
     String password;
 
-    public SRAHelper(String sraUrl, String username, String password)
-    {
+    private BuildProgressLogger buildProgressLogger;
+    private Boolean logRESTCalls = false;
+
+    public SRAHelper(String sraUrl, String username, String password) {
         this.sraUrl = sraUrl;
         this.username = username;
         this.password = password;
@@ -68,13 +69,21 @@ public class SRAHelper
     public String getComponentId(String componentName)
             throws Exception {
         String result = null;
-        URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("component").queryParam("name", componentName)
-                .build();
+        URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("component").
+                queryParam("name", componentName).build();
         String componentContent = executeJSONGet(uri);
         JSONArray components = new JSONArray(componentContent);
         JSONObject component = components.getJSONObject(0);
         result = component.getString("id");
         return result;
+    }
+
+    public void setBuildProgressLogger(BuildProgressLogger buildProgressLogger) {
+        this.buildProgressLogger = buildProgressLogger;
+    }
+
+    public void setLogRestCalls(Boolean logRESTCalls) {
+        this.logRESTCalls = logRESTCalls;
     }
 
     public String getComponentRepositoryId(String componentName)
@@ -83,7 +92,6 @@ public class SRAHelper
 
         URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("component").path(componentName)
                 .build();
-
         String componentContent = executeJSONGet(uri);
 
         JSONArray properties = new JSONObject(componentContent).getJSONArray("properties");
@@ -106,7 +114,6 @@ public class SRAHelper
     public int getNumComponentVersions(String componentId)
             throws Exception {
         URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("component").path(componentId).build();
-
         String componentJson = executeJSONGet(uri);
         JSONObject componentObj = new JSONObject(componentJson);
         int totalRecords = componentObj.getInt("componentVersionCount");
@@ -119,7 +126,8 @@ public class SRAHelper
 
         int numVersions = getNumComponentVersions(componentId);
 
-        System.out.println("##teamcity[message text='Found " + String.valueOf(numVersions) + " versions' status='NORMAL']");
+        buildProgressLogger.progressMessage("Checking " + numVersions + " version(s) to see if version \"" +
+                versionName + "\" exists");
         URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("component").path(componentId)
                 .path("versionsPaged").queryParam("inactive", "true")
                 .queryParam("orderField", "dateCreated")
@@ -131,6 +139,11 @@ public class SRAHelper
         String versionsJson = executeJSONGet(uri);
         JSONObject versionsObj = new JSONObject(versionsJson);
         int totalRecords = versionsObj.getInt("totalRecords");
+        if (totalRecords != numVersions) {
+            buildProgressLogger.progressMessage("Warning, found " + totalRecords + " version(s) in component not " + numVersions);
+        }
+        buildProgressLogger.progressMessage("Checking if version name \"" + versionName + "\" exists in component: " + componentId);
+        boolean foundVersion = false;
         if (totalRecords > 0) {
             JSONArray verArray = versionsObj.getJSONArray("records");
             // see if we can match exact version name
@@ -138,8 +151,13 @@ public class SRAHelper
                 JSONObject verObj = verArray.getJSONObject(i);
                 if (verObj.getString("name").equals(versionName)) {
                     result = verObj.getString("id");
+                    buildProgressLogger.progressMessage("Found version...");
+                    foundVersion = true;
                     break;
                 }
+            }
+            if (!foundVersion) {
+                buildProgressLogger.progressMessage("Could not find version...");
             }
         }
 
@@ -147,13 +165,11 @@ public class SRAHelper
     }
 
     public String getComponentVersionPropsheetId(String verId)
-        throws Exception {
+            throws Exception {
         URI uri = UriBuilder.fromPath(getSraUrl()).path("rest").path("deploy").path("version")
                 .path(verId).build();
         String result = null;
-
         String versionContent = executeJSONGet(uri);
-
         JSONArray propSheets = new JSONObject(versionContent).getJSONArray("propSheets");
         if (propSheets != null) {
             JSONObject propertyJson = propSheets.getJSONObject(0);
@@ -163,7 +179,7 @@ public class SRAHelper
     }
 
     public String createProcessRequest(String componentName, String versionName, String jsonProperties,
-                                     String appName, String envName, String procName)
+                                       String appName, String envName, String procName)
             throws Exception {
         URI uri = UriBuilder.fromPath(getSraUrl()).path("cli").path("applicationProcessRequest")
                 .path("request").build();
@@ -175,8 +191,143 @@ public class SRAHelper
                         ",\"versions\":[{\"version\":\"" + versionName +
                         "\",\"component\":\"" + componentName + "\"}]}";
 
-        String jsonOut = executeJSONPut(uri,jsonIn);
+        String jsonOut = executeJSONPut(uri, jsonIn);
         return jsonOut; // return string so we can log it
+    }
+
+    public List<String> getAllComponentNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("component")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getComponentVersionNames(String componentName) throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("component")
+                .path(componentName)
+                .path("versions")
+                .path("false")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getAllApplicationNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("application")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getApplicationEnvironmentNames(String applicationName) throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("application")
+                .path(applicationName)
+                .path("environments")
+                .path("false")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getAllEnvironmentNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("globalEnvironment")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getApplicationEnvironmentNamesOrAllEnvironmentNames(String applicationName) throws Exception {
+        if (applicationName == null || applicationName.trim().isEmpty()) {
+            return getAllEnvironmentNames();
+        } else {
+            return getApplicationEnvironmentNames(applicationName);
+        }
+    }
+
+    public List<String> getApplicationProcessNames(String applicationName) throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("application")
+                .path(applicationName)
+                .path("processes")
+                .path("false")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getAllApplicationProcessNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("applicationProcess")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getApplicationProcessNamesOrAllApplicationProcessNames(String applicationName) throws Exception {
+        if (applicationName == null || applicationName.trim().isEmpty()) {
+            return getAllApplicationProcessNames();
+        } else {
+            return getApplicationProcessNames(applicationName);
+        }
+    }
+
+    public List<String> getApplicationComponentNames(String applicationName) throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("application")
+                .path(applicationName)
+                .path("components")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getApplicationComponentNamesOrAllComponentNames(String applicationName) throws Exception {
+        if (applicationName == null || applicationName.trim().isEmpty()) {
+            return getAllComponentNames();
+        } else {
+            return getApplicationComponentNames(applicationName);
+        }
+    }
+
+    public List<String> getAllVersionStatusNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("deploy")
+                .path("status")
+                .path("versionStatuses")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getAllResourceNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("resource")
+                .path("resource")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
+    }
+
+    public List<String> getAllGlobalProcessNames() throws Exception {
+        URI uri = UriBuilder.fromPath(getSraUrl())
+                .path("rest")
+                .path("process")
+                .path("false")
+                .build();
+        return executeGetRequestAndGetObjectNamesFromJsonArray(uri);
     }
 
     public void verifyConnection() throws Exception {
@@ -215,6 +366,7 @@ public class SRAHelper
         }
 
         GetMethod method = new GetMethod(uri.toString());
+        if (logRESTCalls) buildProgressLogger.progressMessage("Executing GET Request: " + uri.toString());
         setDirectSsoInteractionHeader(method);
         try {
             HttpClientParams params = httpClient.getParams();
@@ -222,27 +374,24 @@ public class SRAHelper
 
             UsernamePasswordCredentials clientCredentials = new UsernamePasswordCredentials(getUsername(), getPassword());
             httpClient.getState().setCredentials(AuthScope.ANY, clientCredentials);
-
             int responseCode = httpClient.executeMethod(method);
             //if (responseCode < 200 || responseCode < 300) {
             if (responseCode == 401) {
-                throw new Exception("Error connecting to Serena DA: Invalid user and/or password");
-            }
-            else if (responseCode != 200) {
-                throw new Exception("Error connecting to Serena DA: " + responseCode);
-            }
-            else {
+                throw new Exception("Error connecting to Micro Focus DA: Invalid user and/or password");
+            } else if (responseCode != 200) {
+                throw new Exception("Error connecting to Micro Focus DA: " + responseCode);
+            } else {
                 result = method.getResponseBodyAsString();
             }
-        }
-        finally {
+        } finally {
             method.releaseConnection();
         }
 
+        if (logRESTCalls) buildProgressLogger.progressMessage("Response: " + result);
         return result;
     }
 
-    public String executeJSONPost(URI uri) throws Exception {
+    public String executeJSONPost(URI uri, String postContents) throws Exception {
         String result = null;
         HttpClient httpClient = new HttpClient();
 
@@ -252,8 +401,15 @@ public class SRAHelper
             Protocol.registerProtocol("https", https);
         }
 
+        if (postContents == null || postContents.length() == 0) postContents = "{}";
         PostMethod method = new PostMethod(uri.toString());
+        if (logRESTCalls) buildProgressLogger.progressMessage("Executing POST Request: " + uri.toString() +
+                "with body: " + postContents);
+
         setDirectSsoInteractionHeader(method);
+        method.setRequestBody(postContents);
+        method.setRequestHeader("Content-Type", "application/json");
+        method.setRequestHeader("charset", "utf-8");
         method.setRequestHeader("charset", "utf-8");
         try {
             HttpClientParams params = httpClient.getParams();
@@ -264,20 +420,18 @@ public class SRAHelper
 
             int responseCode = httpClient.executeMethod(method);
 
-            if (responseCode != 200 ) {
-                throw new Exception("Serena DA returned error code: " + responseCode);
-            }
-            else {
+            if (responseCode != 200) {
+                throw new Exception("Micro Focus DA returned error code: " + responseCode);
+            } else {
                 result = method.getResponseBodyAsString();
             }
-        }
-        catch (Exception e) {
-            throw new Exception("Error connecting to Serena DA: " + e.getMessage());
-        }
-        finally {
+        } catch (Exception e) {
+            throw new Exception("Error connecting to Micro Focus DA: " + e.getMessage());
+        } finally {
             method.releaseConnection();
         }
 
+        if (logRESTCalls) buildProgressLogger.progressMessage("Response: " + result);
         return result;
     }
 
@@ -291,7 +445,10 @@ public class SRAHelper
             Protocol.registerProtocol("https", https);
         }
 
+        if (putContents == null || putContents.length() == 0) putContents = "{}";
         PutMethod method = new PutMethod(uri.toString());
+        if (logRESTCalls) buildProgressLogger.progressMessage("Executing PUT Request: " + uri.toString() +
+                "with body: " + putContents);
         setDirectSsoInteractionHeader(method);
         method.setRequestBody(putContents);
         method.setRequestHeader("Content-Type", "application/json");
@@ -307,19 +464,17 @@ public class SRAHelper
 
             //if (responseCode < 200 || responseCode < 300) {
             if (responseCode != 200 && responseCode != 204) {
-                throw new Exception("Serena DA returned error code: " + responseCode);
-            }
-            else {
+                throw new Exception("Micro Focus DA returned error code: " + responseCode);
+            } else {
                 result = method.getResponseBodyAsString();
             }
-        }
-        catch (Exception e) {
-            throw new Exception("Error connecting to Serena DA: " + e.getMessage());
-        }
-        finally {
+        } catch (Exception e) {
+            throw new Exception("Error connecting to Micro Focus DA: " + e.getMessage());
+        } finally {
             method.releaseConnection();
         }
 
+        if (logRESTCalls) buildProgressLogger.progressMessage("Response: " + result);
         return result;
     }
 
@@ -328,12 +483,11 @@ public class SRAHelper
     }
 
     public ClientPathEntry[] mergePathEntriesWithCurrentChangeSet(ClientPathEntry[] pathEntries,
-                                                                           ClientChangeSet changeSet) {
+                                                                  ClientChangeSet changeSet) {
         ClientPathEntry[] result;
         if (changeSet == null) {
             result = pathEntries;
-        }
-        else {
+        } else {
             List<ClientPathEntry> fullPathEntries = new ArrayList<ClientPathEntry>();
             Set<String> updatedPaths = new HashSet<String>();
             for (ClientPathEntry pathEntry : pathEntries) {
@@ -354,7 +508,19 @@ public class SRAHelper
     }
 
     public boolean isNotEmpty(String str) {
-        return  (str != null && str.trim().length() > 0);
+        return (str != null && str.trim().length() > 0);
+    }
+
+    private List<String> executeGetRequestAndGetObjectNamesFromJsonArray(URI uri) throws Exception {
+        List<String> names = new ArrayList<>();
+        String jsonArray = executeJSONGet(uri);
+        JSONArray objectArray = new JSONArray(jsonArray);
+        for (int i = 0; i < objectArray.length(); i++) {
+            JSONObject object = objectArray.getJSONObject(i);
+            String name = object.getString("name");
+            names.add(name);
+        }
+        return names;
     }
 
 }