+++ SerenaDeploy-agent/src/main/java/com/serena/SerenaDeployBuildRunner.java
@@ -1,33 +1,26 @@
-
 package com.serena;
-
 
 import jetbrains.buildServer.RunBuildException;
 import jetbrains.buildServer.agent.*;
-import org.jetbrains.annotations.*;
+import org.jetbrains.annotations.NotNull;
 
-public class SerenaDeployBuildRunner implements AgentBuildRunner, AgentBuildRunnerInfo
-{
+public class SerenaDeployBuildRunner implements AgentBuildRunner, AgentBuildRunnerInfo {
     @NotNull
-    public BuildProcess createBuildProcess(@NotNull final AgentRunningBuild runningBuild, @NotNull final BuildRunnerContext context) throws RunBuildException
-    {
+    public BuildProcess createBuildProcess(@NotNull final AgentRunningBuild runningBuild, @NotNull final BuildRunnerContext context) throws RunBuildException {
         return new SerenaDeployBuildProcess(runningBuild, context);
     }
 
     @NotNull
-    public AgentBuildRunnerInfo getRunnerInfo()
-    {
+    public AgentBuildRunnerInfo getRunnerInfo() {
         return this;
     }
 
     @NotNull
-    public String getType()
-    {
-        return PluginConstants.RunType;
+    public String getType() {
+        return SerenaDeployPluginConstants.RUN_TYPE;
     }
 
-    public boolean canRun(@NotNull final BuildAgentConfiguration agentConfiguration)
-    {
+    public boolean canRun(@NotNull final BuildAgentConfiguration agentConfiguration) {
         return true;
     }
 }