+++ SerenaDeploy-common/src/main/java/com/serena/PluginConstants.java
@@ -1,7 +1,0 @@
-package com.serena;
-
-public class PluginConstants {
-    public static String DisplayName = "Serena Deploy";
-    public static String RunType = "SerenaDeploy";
-    public static String Description = "This plugin integrates TeamCity with Micro Focus | Serena Deployment Automation.";
-}