+++ SerenaDeploy-common/src/main/java/com/serena/SerenaDeployPluginConstants.java
@@ -1,0 +1,8 @@
+package com.serena;
+
+public class SerenaDeployPluginConstants {
+    public static String NAME = "serenaDeploy";
+    public static String DISPLAY_NAME = "Serena Deploy";
+    public static String RUN_TYPE = "SerenaDeploy";
+    public static String DESCRIPTION = "This plugin integrates TeamCity with Micro Focus | Serena Deployment Automation.";
+}